requirements.txt
Flask~=2.0.2
requests~=2.27.1
pytest~=6.2.5
setuptools~=60.5.0
python-dotenv~=0.19.2